auto_pageant_loader/requirements.txt
PyAutoGUI==0.9.54
python-dotenv==1.1.0
